refactor(epicenter): consolidate ProviderContext types into provider.shared.ts
Unify ProviderContext across Node.js and browser environments:
- Create single source of truth in provider.shared.ts
- Use discriminated union: paths is defined in Node.js, undefined in browser
- Update markdown-provider to check for paths before destructuring
- Re-export from platform-specific files for backward compatibility
